update activity donate
